requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydantic>=2.12.0,<3.0.0
openai==1.35.14
moviepy==2.2.1
Pillow==11.3.0
numpy==2.4.4
rich==15.0.0
